FIX: switch to double-precision float
git-svn-id: https://www.nitrc.org/svn/cmtk/trunk@3139 42a5c34f-2066-0410-bec5-ba365beb4995

+++ core/libs/Segmentation/cmtkLabelCombinationMultiClassSTAPLE.h
@@ -60,7 +60,7 @@ public:
   typedef LabelCombinationMultiClassSTAPLE Self;
 
   /// Real value type for internal computations.
-  typedef float RealValueType;
+  typedef double RealValueType;
 
   /// Confusion matrix type.
   typedef Matrix2D<RealValueType> ConfusionMatrixType;
